half precision training

+++ model.py
@@ -6,7 +6,9 @@ class AutoFocus(nn.Module):
     def __init__(self):
         super().__init__()
         self.conv_block_1 = nn.Sequential(
-            nn.Conv2d(1, 32, 11, padding=(5, 5)), nn.LeakyReLU(), nn.MaxPool2d(5),
+            nn.Conv2d(1, 32, 11, padding=(5, 5)),
+            nn.LeakyReLU(),
+            nn.MaxPool2d(5),
         )
         self.conv_block_2 = nn.Sequential(
             nn.Conv2d(32, 32, 11, padding=(5, 5)),
@@ -19,7 +21,7 @@ class AutoFocus(nn.Module):
         )
 
     def forward(self, x: torch.Tensor):
-        x = x.float()
+        x = x.half()  # u8 -> f16
         x = self.conv_block_1(x)
         x = self.conv_block_2(x)
         x = self.fc_block(x)
